Move functions out of common
