+++ scripts/get_collaterals.py
@@ -10,7 +10,56 @@ import argparse
 import csv
 import sys
 from collections import defaultdict
-from common import get_web3_connection, get_deposit_events, get_miner_collateral
+from common import get_web3_connection, get_miner_collateral
+
+
+def get_deposit_events(w3, contract_address, block_num_low, block_num_high):
+    """Fetch all Deposit events emitted by the Collateral contract within a block range.
+
+    Args:
+        w3 (Web3): Web3 instance to use for blockchain interaction
+        contract_address (str): The address of the deployed Collateral contract
+        block_num_low (int): The starting block number (inclusive)
+        block_num_high (int): The ending block number (inclusive)
+
+    Returns:
+        list[DepositEvent]: List of Deposit events
+    """
+    contract_abi = load_contract_abi()
+
+    contract = w3.eth.contract(address=contract_address, abi=contract_abi)
+
+    checksum_address = w3.to_checksum_address(contract_address)
+
+    event_signature = "Deposit(address,uint256)"
+    event_topic = w3.keccak(text=event_signature).hex()
+
+    filter_params = {
+        "fromBlock": hex(block_num_low),
+        "toBlock": hex(block_num_high),
+        "address": checksum_address,
+        "topics": [event_topic]
+    }
+
+    logs = w3.eth.get_logs(filter_params)
+
+    formatted_events = []
+    for log in logs:
+        account_address = "0x" + log["topics"][1].hex()[-40:]
+        account = w3.to_checksum_address(account_address)
+
+        decoded_event = contract.events.Deposit().process_log(log)
+
+        formatted_events.append(
+            DepositEvent(
+                account=account,
+                amount=decoded_event['args']['amount'],
+                block_number=log["blockNumber"],
+                transaction_hash=log["transactionHash"].hex(),
+            )
+        )
+
+    return formatted_events
 
 
 def main():
